back/requirements.txt
fastapi
uvicorn
pydantic-settings
psycopg
psycopg_pool
pyjwt
pyjwt[crypto]
requests
